requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
pypdf2>=3.0.1
beautifulsoup4>=4.12.0
markdown>=3.5.0
lxml>=5.1.0
pydantic>=2.5.0
rich>=13.7.0
typer>=0.9.0
pytest>=7.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

